Početak rada WebApi
Početak postavljanja kontrolera za završnu aplikaciju, promjena sql baze na stariju inačicu

+++ csharpapp/LjetniRadAplikacija/LjetnaAplikacija/Proizvod.cs
@@ -10,7 +10,7 @@ namespace LjetnaAplikacija
     {
         public string Naziv { get; set; }
         public string Opis { get; set; }
-        public int Cijena { get; set; }
+        public decimal Cijena { get; set; }
         public int Kolicina { get; set; }
         public bool Dostupnost { get; set; }
         public List<Proizvod> Proizvodi { get; set; }
